requirements.txt
fastapi==0.71.0
uvicorn==0.16.0
requests~=2.27.1
pandas~=1.4.2
gspread~=5.3.2
python-crontab==2.6.0
pyairtable
